# api: add role_name and is_learner to the returned peer info