--- a/tests/pdctl/region/region_test.go
+++ b/tests/pdctl/region/region_test.go
@@ -147,31 +147,31 @@ func (s *regionTestSuite) TestRegion(c *C) {
 		args := append([]string{"-u", pdAddr}, testCase.args...)
 		output, e := pdctl.ExecuteCommand(cmd, args...)
 		c.Assert(e, IsNil)
-		regionsInfo := api.RegionsInfo{}
-		c.Assert(json.Unmarshal(output, &regionsInfo), IsNil)
-		pdctl.CheckRegionsInfo(c, regionsInfo, testCase.expect)
+		regions := &api.RegionsInfo{}
+		c.Assert(json.Unmarshal(output, regions), IsNil)
+		pdctl.CheckRegionsInfo(c, regions, testCase.expect)
 	}
 
 	var testRegionCases = []struct {
 		args   []string
-		expect *api.RegionInfo
+		expect *core.RegionInfo
 	}{
 		// region <region_id> command
-		{[]string{"region", "1"}, api.NewRegionInfo(leaderServer.GetRegionInfoByID(1))},
+		{[]string{"region", "1"}, leaderServer.GetRegionInfoByID(1)},
 		// region key --format=raw <key> command
-		{[]string{"region", "key", "--format=raw", "b"}, api.NewRegionInfo(r2)},
+		{[]string{"region", "key", "--format=raw", "b"}, r2},
 		// region key --format=hex <key> command
-		{[]string{"region", "key", "--format=hex", "62"}, api.NewRegionInfo(r2)},
+		{[]string{"region", "key", "--format=hex", "62"}, r2},
 		// issue #2351
-		{[]string{"region", "key", "--format=hex", "622f62"}, api.NewRegionInfo(r2)},
+		{[]string{"region", "key", "--format=hex", "622f62"}, r2},
 	}
 
 	for _, testCase := range testRegionCases {
 		args := append([]string{"-u", pdAddr}, testCase.args...)
 		output, e := pdctl.ExecuteCommand(cmd, args...)
 		c.Assert(e, IsNil)
-		regionInfo := api.RegionInfo{}
-		c.Assert(json.Unmarshal(output, &regionInfo), IsNil)
-		c.Assert(&regionInfo, DeepEquals, testCase.expect)
+		region := &api.RegionInfo{}
+		c.Assert(json.Unmarshal(output, region), IsNil)
+		pdctl.CheckRegionInfo(c, region, testCase.expect)
 	}
 }