Signed-off-by: HunDunDM <hundundm@gmail.com>

<!--
Thank you for working on PD! Please read PD's [CONTRIBUTING](https://github.com/tikv/pd/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed


If you want to open the **Challenge Program** pull request, please use the following template:
https://raw.githubusercontent.com/tikv/.github/master/.github/PULL_REQUEST_TEMPLATE/challenge-program.md
You can use it with query parameters: https://github.com/tikv/pd/compare/master...${you branch}?template=challenge-program.md
-->

### What problem does this PR solve?

<!-- Add the issue link with a summary if it exists. -->

close #3372 [tidb#24293](https://github.com/pingcap/tidb/issues/24293)

### What is changed and how it works?

* api: add `role_name` and `is_learner` to the returned peer info

### Check List

<!-- Remove the items that are not applicable. -->

Tests

<!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)

### Before

```
region 83693
{
  "id": 83693,
  "peers": [
    {
      "id": 83694,
      "store_id": 1
    },
    {
      "id": 83695,
      "store_id": 51
    },
    {
      "id": 83696,
      "store_id": 76
    },
    {
      "id": 104792,
      "store_id": 101,
      "role": 1
    }
  ],
  "leader": {
    "id": 83694,
    "store_id": 1
  },
  ...
}
```

```
SELECT * FROM INFORMATION_SCHEMA.TIKV_REGION_PEERS WHERE REGION_ID=83693;
+-----------+---------+----------+------------+-----------+--------+--------------+
| REGION_ID | PEER_ID | STORE_ID | IS_LEARNER | IS_LEADER | STATUS | DOWN_SECONDS |
+-----------+---------+----------+------------+-----------+--------+--------------+
|     83693 |   83694 |        1 |          0 |         1 | NORMAL |         NULL |
|     83693 |   83695 |       51 |          0 |         0 | NORMAL |         NULL |
|     83693 |   83696 |       76 |          0 |         0 | NORMAL |         NULL |
|     83693 |  104792 |      101 |          0 |         0 | NORMAL |         NULL |
+-----------+---------+----------+------------+-----------+--------+--------------+
```

### After

```
region 83693                                                                                                                                   
{
  "id": 83693,
  "peers": [
    {
      "id": 83694,
      "store_id": 1,
      "role_name": "Voter"
    },
    {
      "id": 83695,
      "store_id": 51,
      "role_name": "Voter"
    },
    {
      "id": 83696,
      "store_id": 76,
      "role_name": "Voter"
    },
    {
      "id": 104792,
      "store_id": 101,
      "role": 1,
      "role_name": "Learner",
      "is_learner": true
    }
  ],
  "leader": {
    "id": 83696,
    "store_id": 76,
    "role_name": "Voter"
  },
  ...
}
```

```
SELECT * FROM INFORMATION_SCHEMA.TIKV_REGION_PEERS WHERE REGION_ID=83693;
+-----------+---------+----------+------------+-----------+--------+--------------+
| REGION_ID | PEER_ID | STORE_ID | IS_LEARNER | IS_LEADER | STATUS | DOWN_SECONDS |
+-----------+---------+----------+------------+-----------+--------+--------------+
|     83693 |   83694 |        1 |          0 |         0 | NORMAL |         NULL |
|     83693 |   83695 |       51 |          0 |         0 | NORMAL |         NULL |
|     83693 |   83696 |       76 |          0 |         1 | NORMAL |         NULL |
|     83693 |  104792 |      101 |          1 |         0 | NORMAL |         NULL |
+-----------+---------+----------+------------+-----------+--------+--------------+
```

Side effects

- Possible performance regression
  - When `/regions` returns a large number of regions

### Before

```
goos: darwin
goarch: amd64
pkg: github.com/tikv/pd/server/api
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
BenchmarkConvertToAPIRegions-4               571           1918032 ns/op
BenchmarkConvertToAPIRegions-4               584           1892309 ns/op
BenchmarkConvertToAPIRegions-4               570           1918750 ns/op
BenchmarkConvertToAPIRegions-4               592           2220998 ns/op
BenchmarkConvertToAPIRegions-4               598           1936396 ns/op
```

### After

```
goos: darwin
goarch: amd64
pkg: github.com/tikv/pd/server/api
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
BenchmarkConvertToAPIRegions-4               368           3293800 ns/op
BenchmarkConvertToAPIRegions-4               375           3128312 ns/op
BenchmarkConvertToAPIRegions-4               368           3121975 ns/op
BenchmarkConvertToAPIRegions-4               374           3114500 ns/op
BenchmarkConvertToAPIRegions-4               373           3083650 ns/op
```

Related changes

- Need to cherry-pick to the release branch

### Release note

<!-- A bugfix or a new feature needs a release note. If there is no need release note, just uncomment the below line. -->

* Fix the bug that the `is_learner` field of the Region Peer in TiDB is incorrect.